工具/aws-s3-uploader/requirements.txt
boto3==1.34.0
python-dotenv==1.0.0
click==8.1.7 